+++ artifex/models/classification/binary_classification/spam_detection/__init__.py
@@ -5,7 +5,7 @@ from datasets import ClassLabel
 
 from ...classification_model import ClassificationModel
 
-from artifex.core import auto_validate_methods, ParsedModelInstructions, track_training_calls
+from artifex.core import auto_validate_methods, ParsedModelInstructions, track_training_calls, ValidationError
 from artifex.config import config
 
 
@@ -90,9 +90,10 @@ class SpamDetection(ClassificationModel):
     
     @track_training_calls
     def train(
-        self, spam_content: list[str], language: str = "english", output_path: Optional[str] = None, 
+        self, spam_content: Optional[list[str]] = None, language: str = "english", output_path: Optional[str] = None, 
         num_samples: int = config.DEFAULT_SYNTHEX_DATAPOINT_NUM, num_epochs: int = 3,
-        device: Optional[int] = None, disable_logging: Optional[bool] = False
+        device: Optional[int] = None, disable_logging: Optional[bool] = False,
+        train_dataset_path: Optional[str] = None
     ) -> TrainOutput:
         f"""
         Overrides `ClassificationModel.train` to remove the `domain` and `classes` arguments and
@@ -112,14 +113,20 @@ class SpamDetection(ClassificationModel):
             TrainOutput: The output of the training process.
         """
         
+        if train_dataset_path is None and spam_content is None:
+            raise ValidationError(
+                message="The `spam_content` parameter is required when `train_dataset_path` is not provided."
+            )
+
         user_instructions = self._parse_user_instructions(
             user_instructions=spam_content,
             language=language
-        )
+        ) if spam_content else None
         
         output: TrainOutput = self._train_pipeline(
             user_instructions=user_instructions, output_path=output_path, num_samples=num_samples, 
-            num_epochs=num_epochs, device=device
+            num_epochs=num_epochs, device=device,
+            train_dataset_path=train_dataset_path
         )
         
         return output